refactor: rename s57_world_generator to marine_charts_to_gazebo_world
Broader name reflects the package's scope beyond S57 charts — it also
uses ETOPO bathymetry and will support OSM data in the future.

Updates all imports, entry points, cache paths, setup metadata, and
test references.
