+++ marine_charts_to_gazebo_world/marine_charts_to_gazebo_world/generate_world.py
@@ -173,7 +173,7 @@ def _write_readme(args, bbox, terrain_info, heightmap_info):
     readme_path = os.path.join(args.output_dir, "README.md")
     with open(readme_path, "w") as f:
         f.write(f"# {args.world_name}\n\n")
-        f.write(f"Generated by `s57_world_generator` on "
+        f.write(f"Generated by `marine_charts_to_gazebo_world` on "
                 f"{datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n\n")
         f.write("## Parameters\n\n")
         f.write(f"- **Bounds**: {bbox.south:.6f}, {bbox.west:.6f}, "
