+++ marine_charts_to_gazebo_world/marine_charts_to_gazebo_world/heightmap.py
