+++ marine_charts_to_gazebo_world/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/marine_charts_to_gazebo_world
+[install]
+install_scripts=$base/lib/marine_charts_to_gazebo_world
