--- s57_world_generator/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/s57_world_generator
-[install]
-install_scripts=$base/lib/s57_world_generator
